+++ convex/_generated/api.d.ts
@@ -13,8 +13,10 @@ import type {
   FilterApi,
   FunctionReference,
 } from "convex/server";
+import type * as auth from "../auth.js";
 import type * as board from "../board.js";
 import type * as crons from "../crons.js";
+import type * as http from "../http.js";
 
 /**
  * A utility for referencing Convex functions in your app's API.
@@ -25,8 +27,10 @@ import type * as crons from "../crons.js";
  * ```
  */
 declare const fullApi: ApiFromModules<{
+  auth: typeof auth;
   board: typeof board;
   crons: typeof crons;
+  http: typeof http;
 }>;
 export declare const api: FilterApi<
   typeof fullApi,